move naive predictors to independent subpackages (#788)
