+++ src/gluonts/evaluation/_base.py
@@ -503,7 +503,7 @@ class Evaluator:
         https://www.m4.unic.ac.cy/wp-content/uploads/2018/03/M4-Competitors-Guide.pdf
         """
         # avoid import error due to circular dependency
-        from gluonts.model.baseline import naive_2
+        from gluonts.model.naive_2 import naive_2
 
         # calculate the forecast of the seasonal naive predictor
         naive_median_fcst = naive_2(
