+++ src/gluonts/model/seasonal_naive/_predictor.py
