+++ test/model/naive_predictors/r_naive_2_inputs.csv
