+++ test/model/naive_predictors/test_r_code_compliance_of_naive_2.py
@@ -18,8 +18,7 @@ import os
 from pathlib import Path
 
 # First-party imports
-from gluonts.dataset.common import Dataset
-from gluonts.model.baseline import naive_2
+from gluonts.model.naive_2 import naive_2
 
 
 # DISCLAIMER:
